% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.functions.R
\name{calc_exterr}
\alias{calc_exterr}
\title{Calc ext err}
\usage{
calc_exterr(df, mass, slope, intercept, N)
}
\arguments{
\item{df}{this will usually be a datafram with heated gas data}

\item{mass}{either 47 or 48}

\item{slope}{york slope from the initial regression}

\item{intercept}{york intercept from the initial regression}

\item{N}{number of heated gases}
}
\description{
Description
}
