% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.functions.R
\name{convert_ARF.D47_to_ARF.Dennis.temp}
\alias{convert_ARF.D47_to_ARF.Dennis.temp}
\title{Dennis Calibration in ARF ref frame, using ARF D47 values}
\usage{
convert_ARF.D47_to_ARF.Dennis.temp(D47)
}
\description{
Dennis Calibration in ARF ref frame, using ARF D47 values
}
