% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.functions.R
\name{convert_ARF.D47_to_ARF.Ghosh.temp}
\alias{convert_ARF.D47_to_ARF.Ghosh.temp}
\title{Ghosh calibration in ARF ref frame, using ARF D47 values}
\usage{
convert_ARF.D47_to_ARF.Ghosh.temp(D47)
}
\description{
Ghosh calibration in ARF ref frame, using ARF D47 values
}
