% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.functions.R
\name{convert_CIT.D47_to_CIT.Ghosh.temp}
\alias{convert_CIT.D47_to_CIT.Ghosh.temp}
\title{Original Ghosh calibration, using CIT reference frame values}
\usage{
convert_CIT.D47_to_CIT.Ghosh.temp(D47)
}
\description{
Original Ghosh calibration, using CIT reference frame values
}
