% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{isoprocessCUBES-package}
\alias{isoprocessCUBES}
\alias{isoprocessCUBES-package}
\title{isoprocessCUBES: CUBES data processor}
\description{
Description
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/isoprocessCUBES}
  \item Report bugs at \url{https://github.com/kopflab/isoprocessCUBES/issues}
}

}
\author{
\strong{Maintainer}: Kathryn Snell \email{kathryn.snell@colorado.edu}

}
\keyword{internal}
