% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.functions.R
\name{lm_eqn}
\alias{lm_eqn}
\title{lm eqn. on plots
form that deals with negatives better, normal linear regression
for just a basic linear regression}
\usage{
lm_eqn(x, y)
}
\description{
lm eqn. on plots
form that deals with negatives better, normal linear regression
for just a basic linear regression
}
