% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.functions.R
\name{york.conf.pred.interval}
\alias{york.conf.pred.interval}
\title{York conf predictio interval}
\usage{
york.conf.pred.interval(reg.df, all.df, mass, slope, intercept, conf.level)
}
\arguments{
\item{reg.df}{dataframe that the york regression is based on - might be HG only or same as all data}

\item{all.df}{all the data of interest for constructing confidence intervals; might be same as reg.df}

\item{mass}{typically either 47 or 48}

\item{slope}{york slope}

\item{intercept}{york intercept}

\item{conf.level}{level of confidence interval, e.g. use 0.95 for 95\% confidence}
}
\description{
Description
}
