% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.functions.R
\name{york.regression}
\alias{york.regression}
\title{York reg general form}
\usage{
york.regression(X, x.err, Y, y.err, error.corr)
}
\description{
script for generating a york regression of a dataset with errors in both x and y
}
